requirements.txt
scapy
requests
beautifulsoup4
python-nmap